moved all rest_service_4 files to new structure project. add dependency to pom.xml and javadoc
