+++ .github/workflows/javadoc.yml
@@ -1,0 +1,33 @@
+name: JavaDoc
+
+on:
+  push:
+    branches: [ '**' ]
+
+permissions:
+  contents: write
+
+jobs:
+  generate-and-publish:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout código fuente
+        uses: actions/checkout@v3
+
+      - name: Configurar JDK 21
+        uses: actions/setup-java@v3
+        with:
+          java-version: '21'
+          distribution: 'temurin'
+          cache: maven
+
+      - name: Generar JavaDoc con Maven
+        run: mvn javadoc:javadoc
+
+      - name: Publicar JavaDoc en GitHub Pages
+        uses: JamesIves/github-pages-deploy-action@v4
+        with:
+          folder: target/site/apidocs
+          branch: gh-pages
+          clean: true
+          single-commit: true
