+++ pom.xml
@@ -8,11 +8,11 @@
 		<version>3.4.4</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.example</groupId>
-	<artifactId>rest-service-4</artifactId>
+	<groupId>com.github.mgrl39</groupId>
+	<artifactId>spring-rest-jsonpatch</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>rest-service-4</name>
-	<description>Demo project for Spring Boot</description>
+	<name>spring-rest-jsonpatch</name>
+	<description>A Spring Boot application demonstrating basic CRUD operations and JSON Patch integration.</description>
 	<url/>
 	<licenses>
 		<license/>
