+++ src/main/java/com/github/mgrl39/springrestjsonpatch/SpringRestJsonPatchApplication.java
@@ -1,13 +1,13 @@
-package com.example.rest_service_4;
+package com.github.mgrl39.springrestjsonpatch;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class RestService4Application {
+public class SpringRestJsonPatchApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(RestService4Application.class, args);
+		SpringApplication.run(SpringRestJsonPatchApplication.class, args);
 	}
 
 }
