+++ src/main/java/com/github/mgrl39/springrestjsonpatch/UserResource.java
@@ -1,4 +1,4 @@
-package com.example.rest_service_4;
+package com.github.mgrl39.springrestjsonpatch;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
