+++ src/test/java/com/github/mgrl39/springrestjsonpatch/RestService4ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.rest_service_4;
+package com.github.mgrl39.springrestjsonpatch;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
